requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
fpdf==1.7.2
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
multidict==6.0.4
openai==0.27.6
Pillow==9.5.0
python-dotenv==1.0.0
requests==2.30.0
sqlparse==0.4.3
tqdm==4.65.0
urllib3==2.0.2
yarl==1.9.2
